requirements.txt
django==2.2
django-model-utils==3.2.0
